+++ packages/blocks-testing/src/vacant_airtable_interface.ts
@@ -65,6 +65,7 @@ export default class VacantAirtableInterface extends AbstractMockAirtableInterfa
             baseData: {
                 id: '',
                 name: '',
+                color: '',
                 activeTableId: '',
                 tableOrder: [],
                 tablesById: {},