+++ packages/sdk/src/models/base.ts
@@ -30,6 +30,7 @@ const WatchableBaseKeys = Object.freeze({
     tables: 'tables' as const,
     collaborators: 'collaborators' as const,
     schema: 'schema' as const,
+    color: 'color' as const,
 });
 
 /**
@@ -116,6 +117,21 @@ class Base extends AbstractModel<BaseData, WatchableBaseKey> {
     get name(): string {
         return this._data.name;
     }
+
+    /**
+     * The color of the base.
+     *
+     * @example
+     * ```js
+     * import {base} from '@airtable/blocks';
+     * import {Box} from '@airtable/blocks/ui';
+     * const exampleBox = <Box backgroundColor={base.color}> This box's background is the same color as the base background</Box>
+     * ```
+     */
+    get color(): string {
+        return this._data.color;
+    }
+
     /**
      * The tables in this base. Can be watched to know when tables are created, deleted, or reordered in the base.
      *
@@ -530,6 +546,10 @@ class Base extends AbstractModel<BaseData, WatchableBaseKey> {
         let didSchemaChange = false;
         if (changedPaths.name) {
             this._onChange(WatchableBaseKeys.name);
+            didSchemaChange = true;
+        }
+        if (changedPaths.color) {
+            this._onChange(WatchableBaseKeys.color);
             didSchemaChange = true;
         }
         if (changedPaths.tableOrder) {