+++ packages/sdk/src/types/base.ts
@@ -19,6 +19,7 @@ export interface ModelChange {
 export interface BaseData {
     id: BaseId;
     name: string;
+    color: string;
     tableOrder: Array<TableId>;
     activeTableId: TableId | null;
     tablesById: ObjectMap<TableId, TableData>;